fix: Ensure that validated new name is unique
